+++ src/slap/matrix.h
@@ -204,73 +204,6 @@ static inline void slap_SetElement(Matrix mat, int row, int col, double val) {
 }
 
 /**
- * @brief Sets all of the elements in a matrix to a single value
- *
- * @param mat Matrix to be modified
- * @param val Value to which each element will be set
- * @return 0 if successful
- */
-enum slap_ErrorCode slap_SetConst(Matrix mat, double val);
-
-/**
- * @brief Copy a matrix to another matrix
- *
- * @param dest a matrix of size (m,n)
- * @param src a matrix of size (n,m)
- * @return 0 if successful
- */
-enum slap_ErrorCode slap_MatrixCopy(Matrix dest, Matrix src);
-
-/**
- * @brief Copy a matrix to another matrix, transposed
- *
- * @param dest a matrix of size (m,n)
- * @param src a matrix of size (n,m)
- * @return 0 if successful
- */
-enum slap_ErrorCode slap_MatrixCopyTranspose(Matrix dest, Matrix src);
-
-/**
- * @brief Copy the data from an array into the matrix, columnwise.
- *
- * @param mat  Matrix with nonzero size
- * @param data Data to be copied into the array. Must have length of at least mat.rows *
- * mat.cols.
- * @return 0 if successful
- */
-enum slap_ErrorCode slap_MatrixCopyFromArray(Matrix mat, const double* data);
-
-/**
- * @brief Set the diagonal elements of the matrix to val, and the rest to zeros.
- *
- * @param mat Square matrix
- * @param val Value for the diagonal elements
- * @return
- */
-enum slap_ErrorCode slap_SetIdentity(Matrix mat, double val);
-
-/**
- * @brief Scale a matrix by a constant factor
- *
- * @param mat Fully initialized matrix of non-zero size. Values will be modified.
- * @param alpha scalar by which to multiply the matrix
- * @return 0 if successsful
- */
-enum slap_ErrorCode slap_ScaleByConst(Matrix mat, double alpha);
-
-/**
- * @brief Set the first n elements of a matrix diagonal from an array
- *
- * If @a len is greater than the minimum dimension, only the minimum dimension will be set.
- * Doesn't touch any of the off-diagonal elements.
- *
- * @param mat Matrix (nrows >= ncols)
- * @param diag Array of length `nrows`.
- * @return
- */
-enum slap_ErrorCode slap_SetDiagonal(Matrix mat, const double* diag, int len);
-
-/**
  * @brief Return the normed difference between 2 matrices of the same size
  *
  * Returns \f$ \sqrt{\sum_{i=0}^{m-1} \sum_{j=0}^{n-1} (A_{ij} - B_{ij})^2 } \f$
